use size_t instead of int32_t to avoid possible loss of data

+++ src/common/inc/tname.h
@@ -25,7 +25,7 @@ void extractTableName(const char *tableId, char *name);
 
 char* extractDBName(const char *tableId, char *name);
 
-int32_t tableIdPrefix(const char* pDb, char* prefix, int32_t len);
+size_t tableIdPrefix(const char* pDb, char* prefix, int32_t len);
 
 void extractTableNameFromToken(SStrToken *pToken, SStrToken* pTable);
